Fixes test warnings where state updates were being triggered after component unmounts
